try to fix upload-bugs

+++ src/app/products/shared/file-upload.service.ts
@@ -12,6 +12,7 @@ import { Product } from '../../models/product.model';
 import { mergeMap } from 'rxjs/operator/mergeMap';
 import { fromPromise } from 'rxjs/observable/fromPromise';
 import { merge } from 'rxjs/observable/merge';
+import { UploadTaskSnapshot } from '@firebase/storage-types';
 
 @Injectable()
 export class FileUploadService {
@@ -44,25 +45,26 @@ export class FileUploadService {
 
     // The storage path
     const path = `product-images/${new Date().getTime()}_${file.name}`;
+    console.log(this.storage.upload);
 
     // The main task
-    this.task = this.storage.upload(path, file);
-
-    // The file's download URL
-    this.downloadURL = this.task.downloadURL();
+    const task = this.storage.upload(path, file);
 
     // Progress monitoring
-    this.percentage = this.task.percentageChanges();
-    this.snapshot = this.task.snapshotChanges();
+    const percentage = task.percentageChanges();
+    const snapshot = task.snapshotChanges();
 
-    return this.saveProduct(data);
+    return this.saveProduct(data, snapshot);
   }
 
-  saveProduct(data: { product: Product; files: FileList }) {
-    return this.snapshot.pipe(
+  saveProduct(
+    data: { product: Product; files: FileList },
+    snapshot: Observable<UploadTaskSnapshot>,
+  ) {
+    return snapshot.pipe(
       concatMap((snap) => {
         console.log(snap);
-        if (snap.downloadURL) {
+        if (snap.downloadURL && snap.bytesTransferred === snap.totalBytes) {
           data.product.imageURLs.push(snap.downloadURL);
           data.product.imageRefs.push(snap.metadata.fullPath);
           this.angularFireDatabase
@@ -76,7 +78,9 @@ export class FileUploadService {
 
   deleteFile(files) {
     files.map((filePath) => {
-      this.storage.ref(filePath).delete();
+      if (this.storage.ref(filePath)) {
+        this.storage.ref(filePath).delete();
+      }
     });
   }
